Rename/reorg map components
