+++ src/lib/maps/SVGLabelLayer/SVGLabelLayer.docs.md
