+++ src/lib/maps/SVGLabelLayer/SVGLabelLayer.svelte
