+++ src/lib/maps/SVGPointLayer/SVGPointLayer.docs.md
